refactor: move main code
